Non-constant declarations inferring `integer` decay
This came out of #92, where a stop-gap solution was to may integer
literals use `i64`. This caused further problems because we do not
support implicit conversion between `i64` and `u64`.

A few notes:
* We want arbitrary-precision integers at compile-time, but we cannot
support them at run-time.
* Integer literals should be usable at run-time without any extra
annotation.

There are at least three approaches worth mentioning.
* Rust has bidirectional inference
* C++ has integer literals be at least of type `int` and otherwise the
  "smallest" signed integer type in which they fit.
* Go gives integer literals a "constant" type and has a default to which
  they implicitly convert if their type is inferred at runtime.

Icarus does not have bidirectional inference, and we generally suspect
this does not fit well with our computation model. Bidirectional
inference means that the type can be affected by expressions far away.
Even if we limit it to the scope of a function, there could be
conflicting expressions. Icarus values having a clear reason for any
inference it makes (c.f., generics/type deduction) as a mechanism to
provide good error messages. Bidirectional inference and constraint
systems afford us the ability to learn that a conflict exists, but not
which value/type is incorrect.

The C++ system has further problems; It cannot actually support
arbitary-precision integers at compile-time without significant changes.
Because a bounded type is chosen for literals, supporting arbitrary
precision would mean departing from the existing definitions of binary operators
at compile-time. This would have significant cascading affects:
Functions templated on an integer might deduce different return types at
runtime or compile-time.

Go's system seems reasonable, and the one we've implemented here, but is not
without problems. The choice of type to which a compile-time `integer`
decays should probably be platform specific. If that decay happens at
compile-time (i.e., a non-constant is evaluated at compile-time) we need
to make decisions about how this works with cross-compilation. It is
also just a bit strang that the type of `3` and `n` given `n := 3` are
not identical. At its core though these problems seem least intrusive,
and the cross-compilation issues are ones we need to solve anyway.

+++ compiler/verify/declaration_test.cc
@@ -82,12 +82,11 @@ TEST(Declaration, DefaultInitNonConstantType) {
 // TODO Default initialization of non-default-initializable type (both const and
 // non-const).
 
-// TODO non-constant binding of compile-time only values.
 TEST(Declaration, InferredSuccess) {
   {
     test::TestModule mod;
     mod.AppendCode(R"(
-    n := 3 as i64
+    n := 3
     )");
     auto qts = mod.context().qual_types(mod.Append<ast::Identifier>("n"));
     EXPECT_THAT(qts, UnorderedElementsAre(
